backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
slowapi>=0.1.9
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1
langgraph>=0.0.50
chromadb>=0.4.0
pypdf>=4.0.0
docx2txt>=0.8
python-multipart>=0.0.6
sentence-transformers>=2.2.0
tiktoken>=0.5.0
pydantic>=2.0.0
